fix tests after class removal
