+++ hoot-services/src/test/java/hoot/services/controllers/export/ExportCommandTest.java
@@ -56,6 +56,7 @@ public class ExportCommandTest {
         exportParams.setTextStatus(false);
         exportParams.setInputType("file");
         exportParams.setOutputType("shp");
+        exportParams.setIncludeHootTags(false);
 
         ExportCommand exportCommand = new ExportCommand(jobId, exportParams, debugLevel, caller, null);
 
@@ -69,8 +70,8 @@ public class ExportCommandTest {
         assertNotNull(exportCommand.getWorkDir());
         assertNotNull(exportCommand.getCommand());
 
-        String expectedCommand = "hoot convert --${DEBUG_LEVEL} -C RemoveReview2Pre.conf " +
-                "${HOOT_OPTIONS} ${INPUT_PATH} ${OUTPUT_PATH} --trans ${TRANSLATION_PATH}";
+        String expectedCommand = "hoot convert --${DEBUG_LEVEL} " +
+                "${HOOT_OPTIONS} ${INPUT_PATH} ${OUTPUT_PATH}";
         assertEquals(expectedCommand, exportCommand.getCommand());
 
         assertTrue(exportCommand.getSubstitutionMap().containsKey("DEBUG_LEVEL"));
@@ -87,4 +88,59 @@ public class ExportCommandTest {
         assertTrue(exportCommand.getSubstitutionMap().containsKey("OUTPUT_PATH"));
         assertEquals(expectedOutputPath, exportCommand.getSubstitutionMap().get("OUTPUT_PATH"));
     }
+
+    @Test
+    public void testExportOSMCommand() {
+        String jobId = UUID.randomUUID().toString();
+        String debugLevel = "error";
+        Class<?> caller = this.getClass();
+
+        ExportParams exportParams = new ExportParams();
+        exportParams.setInput("input");
+        exportParams.setOutputName("output");
+        exportParams.setTextStatus(true);
+        exportParams.setInputType("file");
+        exportParams.setOutputType("shp");
+        exportParams.setIncludeHootTags(false);
+        exportParams.setAppend(false);
+
+        ExportCommand exportCommand = new ExportCommand(jobId, exportParams, debugLevel, caller, null);
+
+        List<String> options = new LinkedList<>();
+        options.add("api.db.email=test@test.com");
+        options.add("remove.elements.visitor.element.criteria=hoot::ReviewRelationCriterion");
+        options.add("remove.elements.visitor.recursive=false");
+        options.add("writer.include.circular.error.tags=false");
+        options.add("writer.text.status=true");
+        options.add("convert.ops=hoot::RemoveElementsVisitor;hoot::DecomposeBuildingRelationsVisitor");
+
+        List<String> hootOptions = new LinkedList<>();
+
+        options.forEach(option -> { hootOptions.add("-D"); hootOptions.add(option); });
+
+        assertEquals(jobId, exportCommand.getJobId());
+        assertEquals(true, exportCommand.getTrackable());
+        assertNotNull(exportCommand.getSubstitutionMap());
+        assertNotNull(exportCommand.getWorkDir());
+        assertNotNull(exportCommand.getCommand());
+
+        String expectedCommand = "hoot convert --${DEBUG_LEVEL} ${HOOT_OPTIONS} ${INPUT_PATH} ${OUTPUT_PATH}";
+        assertEquals(expectedCommand, exportCommand.getCommand());
+
+        assertTrue(exportCommand.getSubstitutionMap().containsKey("DEBUG_LEVEL"));
+        assertEquals(debugLevel, exportCommand.getSubstitutionMap().get("DEBUG_LEVEL"));
+
+        assertTrue(exportCommand.getSubstitutionMap().containsKey("HOOT_OPTIONS"));
+        assertEquals(hootOptions, exportCommand.getSubstitutionMap().get("HOOT_OPTIONS"));
+
+        assertTrue(exportCommand.getSubstitutionMap().containsKey("INPUT_PATH"));
+        assertEquals(exportParams.getInput(), exportCommand.getSubstitutionMap().get("INPUT_PATH"));
+
+        String expectedOutputPath = new File(new File(TEMP_OUTPUT_PATH, jobId),
+                exportParams.getOutputName() + "." + exportParams.getOutputType()).getAbsolutePath();
+
+        assertTrue(exportCommand.getSubstitutionMap().containsKey("OUTPUT_PATH"));
+        assertEquals(expectedOutputPath, exportCommand.getSubstitutionMap().get("OUTPUT_PATH"));
+    }
+
 }
